A bit more modular - rollback commit
