+++ .gitignore
@@ -1,6 +1,7 @@
 __pycache__/
 Lib/
 Scripts/
+Include/
 selenium/
 *.cfg
 *.json
