+++ SpookyBot/commands/fun/is_it_halloween.py
