+++ SpookyBot/commands/reference/crypto.py
