+++ SpookyBot/commands/reference/urban_define.py
