+++ SpookyBot/commands/reference/wolfram_alpha.py
